pyproject.toml
[project]
name = "a2a_agent"
version = "0.1.0"
description = "Naive base environment for a2a_agent project."
authors = [{name="acensia", email="acensia@snu.ac.kr"}]
dependencies = [
    "a2a-sdk>=0.2.8",
    "uvicorn>=0.34.3",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta" 
